Fix polish translation mistake
Both "device_status" and "device_status_summary" should be "Informacje" (Status in english), not "Informacje o telefonie" (phone status in english) just like in
a original english translation. The same string is displayed on both phones and tablets, and it says "phone status" also on tablets only in polish translation.

Change-Id: Icf4ef3a70b15f3964f1d53181755b2577d0e6850